requirements.txt
# pip3 install -r requirements.txt
streamlit
ragstack-ai
PyGithub
watchdog
asyncio
rich